// ui/packages/src/api/generated/filesys/schemas/upsertStorageSettingCommand.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface UpsertStorageSettingCommand {
  /** No comments found. */
  configData: string;
  /** No comments found. */
  enabled?: number;
  /** No comments found. */
  id?: string;
  /** No comments found. */
  platformIdentifier: string;
  /** No comments found. */
  remark?: string;
  /** No comments found. */
  userId: string;
}